# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.UserLogItem:
    properties:
      channel:
        example: 1
        type: integer
      completion_tokens:
        example: 34
        type: integer
      content:
        example: request
        type: string
      created_at:
        example: 1700000000
        type: integer
      elapsed_time:
        example: 200
        type: integer
      id:
        example: 1
        type: integer
      is_stream:
        example: false
        type: boolean
      model_name:
        example: gpt-4o-mini
        type: string
      prompt_tokens:
        example: 12
        type: integer
      quota:
        example: 123
        type: integer
      system_prompt_reset:
        example: false
        type: boolean
      token_name:
        example: default
        type: string
      trace_id:
        example: trace_123
        type: string
      type:
        example: 2
        type: integer
      user_id:
        example: 1
        type: integer
      username:
        example: alice
        type: string
    type: object
  docs.UserLogListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/docs.UserLogItem'
        type: array
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  docs.UserLogStatData:
    properties:
      quota:
        example: 12345
        type: integer
    type: object
  docs.UserLogStatResponse:
    properties:
      data:
        $ref: '#/definitions/docs.UserLogStatData'
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
...
  /api/v1/admin/log:
    delete:
      parameters:
      - description: Target timestamp (unix)
        in: query
        name: target_timestamp
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete history logs (admin)
      tags:
      - admin
    get:
      parameters:
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Log type
        in: query
        name: type
        type: integer
      - description: Start timestamp (unix)
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp (unix)
        in: query
        name: end_timestamp
        type: integer
      - description: Username
        in: query
        name: username
        type: string
      - description: Token name
        in: query
        name: token_name
        type: string
      - description: Model name
        in: query
        name: model_name
        type: string
      - description: Group ID
        in: query
        name: group_id
        type: string
      - description: Channel ID
        in: query
        name: channel
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List logs (admin)
      tags:
      - admin
  /api/v1/admin/log/{id}:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get log by ID (admin)
      tags:
      - admin
  /api/v1/admin/log/options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List log filter options (admin)
      tags:
      - admin
  /api/v1/admin/log/search:
    get:
      parameters:
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search logs (admin)
      tags:
      - admin
  /api/v1/admin/log/stat:
    get:
      parameters:
      - description: Log type
        in: query
        name: type
        type: integer
      - description: Start timestamp (unix)
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp (unix)
        in: query
        name: end_timestamp
        type: integer
      - description: Token name
        in: query
        name: token_name
        type: string
      - description: Username
        in: query
        name: username
        type: string
      - description: Model name
        in: query
        name: model_name
        type: string
      - description: Channel ID
        in: query
        name: channel
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogStatResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log stats (admin)
      tags:
      - admin
...
  /api/v1/public/log:
    get:
      parameters:
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Log type
        in: query
        name: type
        type: integer
      - description: Start timestamp (unix)
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp (unix)
        in: query
        name: end_timestamp
        type: integer
      - description: Token name
        in: query
        name: token_name
        type: string
      - description: Model name
        in: query
        name: model_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user logs
      tags:
      - public
  /api/v1/public/log/{id}:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get current user log by ID
      tags:
      - public
  /api/v1/public/log/options:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user log filter options
      tags:
      - public
  /api/v1/public/log/search:
    get:
      parameters:
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search user logs
      tags:
      - public
  /api/v1/public/log/stat:
    get:
      parameters:
      - description: Log type
        in: query
        name: type
        type: integer
      - description: Start timestamp (unix)
        in: query
        name: start_timestamp
        type: integer
      - description: End timestamp (unix)
        in: query
        name: end_timestamp
        type: integer
      - description: Token name
        in: query
        name: token_name
        type: string
      - description: Model name
        in: query
        name: model_name
        type: string
      - description: Channel ID
        in: query
        name: channel
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.UserLogStatResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log stats for current user
      tags:
      - public
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"